+++ Assets/Scenes/Main.unity
@@ -288,53 +288,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
---- !u!1 &162767900
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 162767901}
-  - component: {fileID: 162767902}
-  m_Layer: 0
-  m_Name: Player
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &162767901
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 162767900}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 8
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &162767902
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 162767900}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 565e6e437f4dad3499add23d8d14ea69, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  scoreText: {fileID: 544723528}
-  remainLinesText: {fileID: 1451838258}
-  comboText: {fileID: 1778629756}
-  rushGageImage: {fileID: 616399820}
 --- !u!1 &175879452
 GameObject:
   m_ObjectHideFlags: 0
@@ -801,7 +754,10 @@ MonoBehaviour:
   foodGenerater: {fileID: 1419647445}
   itemGenerater: {fileID: 947819758}
   lineCursol: {fileID: 1712355795}
-  playerController: {fileID: 162767902}
+  scoreText: {fileID: 544723528}
+  remainLinesText: {fileID: 1451838258}
+  comboText: {fileID: 1778629756}
+  rushGageImage: {fileID: 616399820}
   calorieGageImage: {fileID: 91724116}
 --- !u!4 &658604814
 Transform:
@@ -936,7 +892,6 @@ MonoBehaviour:
   gameManager: {fileID: 658604813}
   foodGenerater: {fileID: 1419647445}
   itemGenerater: {fileID: 947819758}
-  playerController: {fileID: 162767902}
 --- !u!4 &852684521
 Transform:
   m_ObjectHideFlags: 0
